fix make

+++ module_00/ex00/megaphone.cpp
@@ -9,7 +9,7 @@ int main(int argc, char **argv) {
 	for (int i = 1; i < argc; i++) {
 		std::string str = argv[i];
 
-		for (int j = 0; j < str.length(); j++)
+		for (unsigned long j = 0; j < str.length(); j++)
 			std::cout << (char) toupper(str[j]);
 	}
 	std::cout << std::endl;
